# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AdminOnboardingSnapshotResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/MarketplaceEventRecord"
            },
            "title": "Events",
            "type": "array"
          },
          "registrations": {
            "items": {
              "$ref": "#/components/schemas/TargetRegistrationRecord"
            },
            "title": "Registrations",
            "type": "array"
          }
        },
        "title": "AdminOnboardingSnapshotResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketplaceEventIngestRequest": {
        "properties": {
          "container_app_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container App Name"
          },
          "container_app_resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container App Resource Id"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "environment": {
            "default": "prod",
            "title": "Environment",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "event_type": {
            "default": "subscription_purchased",
            "title": "Event Type",
            "type": "string"
          },
          "health_status": {
            "default": "registered",
            "title": "Health Status",
            "type": "string"
          },
          "last_deployed_release": {
            "default": "unknown",
            "title": "Last Deployed Release",
            "type": "string"
          },
          "managed_application_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Application Id"
          },
          "managed_resource_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Resource Group Id"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "subscription_id": {
            "title": "Subscription Id",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          },
          "target_group": {
            "default": "prod",
            "title": "Target Group",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "tier": {
            "default": "standard",
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "tenant_id",
          "subscription_id"
        ],
        "title": "MarketplaceEventIngestRequest",
        "type": "object"
      },
      "MarketplaceEventIngestResponse": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MarketplaceEventStatus"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "required": [
          "event_id",
          "status",
          "message"
        ],
        "title": "MarketplaceEventIngestResponse",
        "type": "object"
      },
      "MarketplaceEventRecord": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "status": {
            "$ref": "#/components/schemas/MarketplaceEventStatus"
          },
          "subscription_id": {
            "title": "Subscription Id",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "event_type",
          "status",
          "message",
          "tenant_id",
          "subscription_id",
          "created_at"
        ],
        "title": "MarketplaceEventRecord",
        "type": "object"
      },
      "MarketplaceEventStatus": {
        "enum": [
          "applied",
          "duplicate",
          "rejected"
        ],
        "title": "MarketplaceEventStatus",
        "type": "string"
      },
...
      "TargetRegistrationRecord": {
        "properties": {
          "container_app_resource_id": {
            "title": "Container App Resource Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "last_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Id"
          },
          "managed_application_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Application Id"
          },
          "managed_resource_group_id": {
            "title": "Managed Resource Group Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "subscription_id": {
            "title": "Subscription Id",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "target_id",
          "tenant_id",
          "subscription_id",
          "managed_resource_group_id",
          "container_app_resource_id",
          "display_name",
          "created_at",
          "updated_at"
        ],
        "title": "TargetRegistrationRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "MAPPO multi-tenant deployment control plane API",
    "title": "MAPPO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/onboarding": {
      "get": {
        "operationId": "get_onboarding_snapshot_api_v1_admin_onboarding_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 250,
              "minimum": 1,
              "title": "Event Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOnboardingSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Onboarding Snapshot",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/onboarding/events": {
      "post": {
        "operationId": "ingest_marketplace_event_api_v1_admin_onboarding_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-mappo-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Mappo-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketplaceEventIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceEventIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Marketplace Event",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}